chore: modularized world js files
